pyproject.toml
[project]
name = "meshtastic-ui-ha"
version = "0.2.0"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-homeassistant-custom-component",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
